+++ src/pluginsettings.cpp
@@ -12,6 +12,7 @@ void StoermelderSettings::saveToJson() {
 	json_object_set(settingsJ, "mbV1zoom", json_real(mbV1zoom));
 	json_object_set(settingsJ, "mbV1sort", json_integer(mbV1sort));
 	json_object_set(settingsJ, "mbV1hideBrands", json_boolean(mbV1hideBrands));
+	json_object_set(settingsJ, "mbV1searchDescriptions", json_boolean(mbV1searchDescriptions));
 
 	json_object_set(settingsJ, "midiLoopbackDriverEnabled", json_boolean(midiLoopbackDriverEnabled));
 
@@ -58,6 +59,8 @@ void StoermelderSettings::readFromJson() {
 	if (mbV1sortJ) mbV1sort = json_integer_value(mbV1sortJ);
 	json_t* mbV1hideBrandsJ = json_object_get(settingsJ, "mbV1hideBrands");
 	if (mbV1hideBrandsJ) mbV1hideBrands = json_boolean_value(mbV1hideBrandsJ);
+    json_t* mbV1searchDescriptionsJ = json_object_get(settingsJ, "mbV1searchDescriptions");
+    if (mbV1searchDescriptionsJ) mbV1searchDescriptions = json_boolean_value(mbV1searchDescriptionsJ);
 
 	json_t* midiLoopbackDriverEnabledJ = json_object_get(settingsJ, "midiLoopbackDriverEnabled");
 	if (midiLoopbackDriverEnabledJ) midiLoopbackDriverEnabled = json_boolean_value(midiLoopbackDriverEnabledJ);