+++ src/pluginsettings.hpp
@@ -8,6 +8,7 @@ struct StoermelderSettings {
 	float mbV1zoom = 0.85f;
 	int mbV1sort = 0;
 	bool mbV1hideBrands = false;
+	bool mbV1searchDescriptions = false;
 
 	bool midiLoopbackDriverEnabled = false;